+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add block list event.
+  - Conn: `BlockListConn`
+  - Dns: `BlockListDns`
+
 ### Changed
 
 - Modified `FtpBruteForce`, `LdapBruteForce`, `RdpBruteForce` events to align